+++ src/Exception/InvalidHeader.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Stadly\Http\Exception;
+
+use RuntimeException;
+
+/**
+ * Exception thrown when header is invalid.
+ */
+final class InvalidHeader extends RuntimeException
+{
+}