+++ src/Header/Request/IfNoneMatch.php
@@ -4,8 +4,11 @@ declare(strict_types=1);
 
 namespace Stadly\Http\Header\Request;
 
+use Stadly\Http\Exception\InvalidHeader;
 use Stadly\Http\Header\Value\EntityTag\EntityTag;
 use Stadly\Http\Header\Value\EntityTag\EntityTagSet;
+use Stadly\Http\Utilities\Rfc7230;
+use Stadly\Http\Utilities\Rfc7232;
 
 /**
  * Class for handling the HTTP header field If-None-Match.
@@ -34,9 +37,15 @@ final class IfNoneMatch implements Header
      *
      * @param string $value Header value.
      * @return self Header generated based on the value.
+     * @throws InvalidHeader If the header value is invalid.
      */
     public static function fromValue(string $value): self
     {
+        $regEx = '{^' . Rfc7230::hashRule(Rfc7232::ENTITY_TAG, 1) . '$}';
+        if (utf8_decode($value) !== $value || preg_match($regEx, $value) !== 1) {
+            throw new InvalidHeader('Invalid header value: ' . $value);
+        }
+
         return new self(EntityTagSet::fromString($value));
     }